ver 0.3 change structure of project, update simulation, create IServices

+++ GenericHostSimulation_lab3/Services/SimulationOptions.cs
@@ -2,7 +2,7 @@
 
 public class SimulationOptions
 {
-    public int DurationSeconds { get; set; }
+    public long DurationSeconds { get; set; }
     public int ThinkingTimeMin { get; set; }
     public int ThinkingTimeMax { get; set; }
     public int EatingTimeMin { get; set; }
